- patch via koitsu to fix the formatting of time_t in *printf().   we now use IRCD_TIME_FMT in *printf() to get the correct formatting.
git-svn-id: http://svn.ratbox.org/svnroot/trunk@16263 b93f080a-4cfa-0310-a19a-80a68f531ef9

+++ src/ircd.c
@@ -284,8 +284,8 @@ set_time(void)
 	if(newtime.tv_sec < CurrentTime)
 	{
 		ircsnprintf(to_send, sizeof(to_send), 
-			   "System clock is running backwards - (%lu < %lu)",
-			   (unsigned long) newtime.tv_sec, (unsigned long) CurrentTime);
+			   "System clock is running backwards - (%ld < %ld)",
+			   newtime.tv_sec, CurrentTime);
 		report_error(to_send, me.name, me.name, 0);
 		set_back_events(CurrentTime - newtime.tv_sec);
 	}
